OSD: Fix edit bar positioning
- use the correct cast to call the correct method
- fixes #13534

+++ mythtv/libs/libmythtv/osd.cpp
@@ -365,7 +365,7 @@ void OSD::SetValues(const QString &Window, const QHash<QString,float> &Map,
         MythUIEditBar *edit = dynamic_cast<MythUIEditBar *> (win->GetChild("editbar"));
         if (edit)
         {
-            edit->SetEditPosition(static_cast<long long>(Map.value("position")));
+            edit->SetEditPosition(static_cast<double>(Map.value("position")));
             found = true;
         }
     }
